requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.4.0
oauthlib==3.0.2
parsel==1.5.1
pathlib2==2.3.4
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
queuelib==1.5.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
Scrapy==1.7.3
service-identity==18.1.0
six==1.12.0
Twisted==19.2.1
uritemplate==3.0.0
urllib3==1.25.3
w3lib==1.20.0
zope.interface==4.6.0
